Fix issue #3, convert server.js to TypeScript, server.ts.
The biggest problem was getting the import statements correct, for some
unknown reason I had to specifically use:

  import * as WebSocket from 'ws';

If I used:

  import * as ws from 'ws';

I ran into a compile time bug that
ws_server.on('connection', (client: WebSocket) ...
was wrong argument type:

  $ tsc
  server.ts(79,28): error TS2345: Argument of type '(client: WebSocket) => void' is not assignable to parameter of type '() => void'.
  server.ts(85,10): error TS2339: Property 'on' does not exist on type 'WebSocket'.
  server.ts(91,10): error TS2339: Property 'on' does not exist on type 'WebSocket'.

The compiler seems to not understand that WebSocket was a type and instead
was interpreting it as a namespace in the statements. I say this because
there are two definitions for WebSocket in
node_modules/@types/ws/websocket/index.d.ts.

One as a class:

  declare class WebSocket extends events.EventEmitter {
      static CONNECTING: number;
      ...

And the other as a namespace:

  declare namespace WebSocket {
      ...


Anyway, that is my guess as to the problem, maybe someday I'll find
the real reason.